requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
flask==3.0.2
flask-cors==4.0.0
flask-marshmallow==1.2.0
flask-sqlalchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
jinja2==3.1.3
markupsafe==2.1.5
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
packaging==23.2
pymysql==1.1.1
python-dotenv==1.0.1
sqlalchemy==2.0.25
typing-extensions==4.9.0
waitress==3.0.0
werkzeug==3.0.1
